% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfrr-package.R
\docType{package}
\name{dfrr-package}
\alias{dfrr-package}
\alias{_PACKAGE}
\title{dfrr: Dichotomized Functional Response Regression}
\description{
Implementing Function-on-Scalar Regression model in which the response function is dichotomized and observed sparsely. This package provides smooth estimations of functional regression coefficients and principal components for the dfrr model.
}
\details{
Implementing Function-on-Scalar Regression model in which the response function is dichotomized and observed sparsely. This package provides smooth estimations of functional regression coefficients and principal components for the dfrr model.
The main function in the dfrr-package is \link{dfrr}().
}
\examples{
set.seed(2000)
N<-50;M<-24
X<-rnorm(N,mean=0)
time<-seq(0,1,length.out=M)
Y<-simulate.simple.dfrr(beta0=function(t){cos(pi*t+pi)},
                        beta1=function(t){2*t},
                        X=X,time=time)

\donttest{dfrr_fit<-dfrr(Y~X,yind=time)}
\dontshow{dfrr_fit<-dfrr(Y~X,yind=time,T_E=3)}

coefs<-coef(dfrr_fit)
  plot(coefs)

fitteds<-fitted(dfrr_fit)
  plot(fitteds)

resids<-residuals(dfrr_fit)
\donttest{plot(resids)}

fpcs<-fpca(dfrr_fit)
\donttest{plot(fpcs,plot.contour=TRUE,plot.3dsurface = TRUE)}

newdata<-data.frame(X=c(1,0))
  preds<-predict(dfrr_fit,newdata=newdata)
  plot(preds)


newdata<-data.frame(X=c(1,0))
newydata<-data.frame(.obs=rep(1,5),.index=c(0.0,0.1,0.2,0.3,0.7),.value=c(1,1,1,0,0))
preds<-predict(dfrr_fit,newdata=newdata,newydata = newydata)
plot(preds)

}
\references{
Fatemeh Asgari, Alamatsaz Mohammad Hossein, Hayati Saeed (2021).
Dichotomized Functional Response Regression Model.
<http://arxive.org/adress_to_paper>
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asgari-fatemeh/dfrr}
  \item Report bugs at \url{https://github.com/asgari-fatemeh/dfrr/issues}
}

}
\author{
\strong{Maintainer}: Fatemeh Asgari \email{ft.asgari@sci.ui.ac.ir}

Authors:
\itemize{
  \item Saeed Hayati \email{s.hayati@sci.ui.ac.ir} [contributor]
}

}
\keyword{Dichotomized}
\keyword{FDA,}
\keyword{Functional}
\keyword{Functional,}
\keyword{Regression}
