% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfrr.R
\name{dfrr}
\alias{dfrr}
\title{Dichotomized Functional Response Regression}
\usage{
dfrr(
  formula,
  yind = NULL,
  data = NULL,
  ydata = NULL,
  method = c("REML", "ML"),
  rangeval = NULL,
  basis = NULL,
  times_to_evaluate = NULL,
  ...
)
}
\arguments{
\item{formula}{an object of class "\code{\link[stats]{formula}}" (or one that can be coerced to that class with \code{as.formula}:
a symbolic description of the model to be fitted.}

\item{yind}{a vector with length equal to the number of columns of the matrix of functional
responses giving the vector of evaluation points \eqn{(t_1,...,t_{G})}.
If not supplied, \code{yind} is set to \code{1:ncol(<response>)}.}

\item{data}{an (optional) \code{data.frame} containing the covariate data.
the variable terms will be searched from the columns of \code{data},
covariates also can be read from the workspace if it is not available in \code{data}.}

\item{ydata}{an (optional) \code{data.frame} consists of three columns \code{.obs}, \code{.index} and \code{.value},
supplying the functional responses that are not observed on a regular grid.
ydata must be provided if the sampling design is irregular.}

\item{method}{detrmines the estimation method of functional parameters. Defaults to "\code{REML}" estimation.}

\item{rangeval}{an (optional) vector of length two, indicating the lower and upper limit of the domain of
latent functional response. If not specified, it will set by minimum and maximum of \code{yind} or \code{.index} column of \code{ydata}.}

\item{basis}{an (optional) object of class \code{'basisfd'}. Defaults to cubic bspline basis.}

\item{...}{other arguments that can be passed to the inner function \code{AMCEM}.}
}
\description{
Implementing Function-on-Scalar Regression model, in which the response function
 is dichotomized and observed sparsely.
}
\details{
The output is a \code{dfrr}-object, which then can be injected into other methods/functions
 to postprocess the fitted model, including:
 \code{\link{coef.dfrr}},\code{\link{fitted.dfrr}}, \code{\link{basis}}, \code{\link{residuals.dfrr}}, \code{\link{predict.dfrr}},
  \code{\link{fpca}}, \code{\link{summary.dfrr}}, \code{\link{model.matrix.dfrr}},
  \code{\link{plot.coef.dfrr}}, \code{\link{plot.fitted.dfrr}}, \code{\link{plot.residuals.dfrr}},
  \code{\link{plot.predict.dfrr}}, \code{\link{plot.fpca.dfrr}}
}
\examples{
set.seed(2000)
N<-50;M<-24
X<-rnorm(N,mean=0)
time<-seq(0,1,length.out=M)
Y<-simulate.simple.dfrr(beta0=function(t){cos(pi*t+pi)},
                        beta1=function(t){2*t},
                        X=X,time=time)
dfrr_fit<-dfrr(Y~X,yind=time)
plot(dfrr_fit)

##### Fitting dfrr model to the Madras Longitudinal Schizophrenia data
data(madras)
ydata<-data.frame(.obs=madras$id,.index=madras$month,.value=madras$y)
ids<-unique(madras$id)
q<-4
N<-length(ids)
xData<-data.frame(Age=rep(NA,N),Gender=rep(NA,N))
for(i in 1:N){
  dt<-madras[madras$id==ids[i],]
  xData[i,]<-c(dt$age[1],dt$gender[1])
}
rownames(xData)<-ids

madras_dfrr<-dfrr(Y~Age+Gender+Age*Gender, data=xData, ydata=ydata, J=11,T_E=5)
coefs<-coef(madras_dfrr)
plot(coefs)

fpcs<-fpca(madras_dfrr)
plot(fpcs)

}
