% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.dfrr.R
\name{fitted.dfrr}
\alias{fitted.dfrr}
\title{Obtain fitted curves for a dfrr model}
\usage{
\method{fitted}{dfrr}(
  dfrr_fit,
  return.fourier.coefs = NULL,
  return.evaluations = !return.fourier.coefs,
  time_to_evaluate = NULL,
  standardized = NULL,
  unstandardized = !standardized
)
}
\arguments{
\item{dfrr_fit}{a fitted \code{dfrr}-object obtained from invoking the function \code{\link{dfrr}}.}

\item{return.fourier.coefs, return.evaluations}{a \code{boolean} indicating whether the Fourier coefficients of the fitted curves are returned
(\code{return.fourier.coefs=TRUE}), or evaluations of the fitted curves (\code{return.evaluations=TRUE}).
Defaults to \code{return.fourier.coefs=TRUE}.}

\item{time_to_evaluate}{a numeric vector indicating the set of time points for evaluating the fitted latent functions, for the case of \code{return.evaluations=TRUE}.}

\item{standardized, unstandardized}{a \code{boolean} indicating whether stanadrdized/unstandardized fitted latent curves is reported.
Only standardized fitted curves are identifiable, thus the arugment is defaults to \code{standardized=TRUE}.}
}
\description{
Fitted curves refer to the estimations of latent functional response curves.
The results can be either the Fourier coefficients or evaluation of the
fitted functions. See Details.
}
\details{
This function will return either the Fourier coefficients or the evaluation of
fitted curves to the binary sequences. Fourier coefficients which are reported are
based on the a set of basis which can be determined by \code{\link{basis}(dfrr_fit)}.
Thus the evaluation of fitted latent curves on the set of time points specified by vector \code{time},
equals to \code{fitted(dfrr_fit)\%*\%t(\link[fda]{eval.basis}(time,\link{basis}(dfrr_fit)))}.

Consider that the unstandardized estimations are not identifiable. So, it is recommended to
extract and report the standardized estimations.
}
\examples{
set.seed(2000)
N<-50;M<-24
X<-rnorm(N,mean=0)
time<-seq(0,1,length.out=M)
Y<-simulate.simple.dfrr(beta0=function(t){cos(pi*t+pi)},
                        beta1=function(t){2*t},
                        X=X,time=time)
\donttest{dfrr_fit<-dfrr(Y~X,yind=time)}
\dontshow{dfrr_fit<-dfrr(Y~X,yind=time,T_E=3)}
fitteds<-fitted(dfrr_fit)
plot(fitteds)

}
\seealso{
\code{\link{plot.fitted.dfrr}}
}
