% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.matrix.dfrr.R
\name{model.matrix.dfrr}
\alias{model.matrix.dfrr}
\title{Obtain model matrix for a dfrr fit}
\usage{
\method{model.matrix}{dfrr}(object, ...)
}
\arguments{
\item{object}{a \code{dfrr}-object}

\item{...}{dot argument, just for consistency with the generic function}
}
\description{
Obtain model matrix for a dfrr fit
}
