% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dfrr.R
\name{plot.dfrr}
\alias{plot.dfrr}
\title{Plot a dfrr fit}
\usage{
\method{plot}{dfrr}(dfrr_fit)
}
\arguments{
\item{dfrr_fit}{a fitted \code{dfrr}-object obtained from invoking the function \code{\link{dfrr}}.}
}
\description{
Plot the regression coefficients, principal components, kernel function and residuals of a \code{dfrr}-object.
}
\details{
The contour plot of the kernel function is produced if the package \code{\link[ggplot2]{ggplot2}} is installed.
Plotting the 3d surface  of the kernel function is also depends on the package \code{\link[plotly]{plotly}}.
To produce the qq-plot, the package \code{\link[car]{car}} must be installed.
}
\examples{
set.seed(2000)
N<-50;M<-24
X<-rnorm(N,mean=0)
time<-seq(0,1,length.out=M)
Y<-simulate.simple.dfrr(beta0=function(t){cos(pi*t+pi)},
                        beta1=function(t){2*t},
                        X=X,time=time)
dfrr_fit<-dfrr(Y~X,yind=time)
plot(dfrr_fit)

}
