% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fitted.dfrr.R
\name{plot.fitted.dfrr}
\alias{plot.fitted.dfrr}
\title{Plot dfrr fitted latent functions}
\usage{
\method{plot}{fitted.dfrr}(
  fitted.dfrr,
  id = NULL,
  main = id,
  col = "blue",
  lwd = 2,
  lty = "solid",
  cex.circle = 1,
  col.circle = "black",
  ...
)
}
\arguments{
\item{fitted.dfrr}{the output of the function \link{fitted.dfrr}}

\item{id}{a vector of length one or more containing subject ids to plot. Must be matched with
\code{rownames(<response>)} or the \code{.obs} column of \code{ydata}. Defaults to
 all  subject ids.}

\item{main}{a vector of length one or \code{length(id)} containing the title of
plots.}

\item{col, lwd, lty, ...}{graphical parameters passed to \code{\link{plot}}}

\item{cex.circle, col.circle}{size and color of circles and filled circles.}
}
\description{
Plot a  \code{fitted.dfrr} object.
}
\details{
The output is the plot of latent curves over the observed binary sequence.
The binary sequence is illustrated with circles and filled circles for the values
of zero and one, respectively.
}
\examples{
set.seed(2000)
N<-50;M<-24
X<-rnorm(N,mean=0)
time<-seq(0,1,length.out=M)
Y<-simulate.simple.dfrr(beta0=function(t){cos(pi*t+pi)},
                        beta1=function(t){2*t},
                        X=X,time=time)
dfrr_fit<-dfrr(Y~X,yind=time)
fitteds<-fitted(dfrr_fit)
plot(fitteds)

}
