% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.residuals.dfrr.R
\name{qq}
\alias{qq}
\title{qq-plot Generic function}
\usage{
qq(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{extra parameters passed to S3 methods}
}
\description{
This is a generic function for qq() method.
}
