% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.dfrr.R
\name{summary.dfrr}
\alias{summary.dfrr}
\title{Summary for a dfrr fit}
\usage{
\method{summary}{dfrr}(object, ..)
}
\arguments{
\item{object}{a \code{dfrr}-object}

\item{...}{dot argument, just for consistency with the generic function}
}
\description{
Summarise a fitted \code{dfrr}-object. Not implemented.
}
