% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.matrix.dfrr.R
\name{model.matrix.dfrr}
\alias{model.matrix.dfrr}
\title{Obtain model matrix for a dfrr fit}
\usage{
\method{model.matrix}{dfrr}(dfrr_fit)
}
\arguments{
\item{dfrr_fit}{a fitted \code{dfrr}-object obtained from invoking the function \code{\link{dfrr}}.}
}
\description{
Obtain model matrix for a dfrr fit
}
