% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.residuals.dfrr.R
\name{plot.residuals.dfrr}
\alias{plot.residuals.dfrr}
\alias{qq.dfrr}
\title{QQ-plot for dfrr residuals}
\usage{
\method{plot}{residuals.dfrr}(residuals.dfrr, ...)

\method{qq}{dfrr}(dfrr_fit, ...)
}
\arguments{
\item{residuals.dfrr}{a \code{residuals.dfrr}-object.}

\item{...}{graphical parameters passed to \code{car::\link[car]{qqPlot}}}

\item{dfrr_fit}{a fitted \code{dfrr}-object obtained from invoking the function \code{\link{dfrr}}.}
}
\description{
The output gives the qq-plot of estimated measurment error.
}
\examples{
\donttest{
N<-50;M<-24
X<-rnorm(N,mean=0)
time<-seq(0,1,length.out=M)
Y<-simulate.simple.dfrr(beta0=function(t){cos(pi*t+pi)},
                        beta1=function(t){2*t},
                        X=X,time=time)
\donttest{dfrr_fit<-dfrr(Y~X,yind=time)}
\dontshow{dfrr_fit<-dfrr(Y~X,yind=time,T_E=3)}
resid<-residuals(dfrr_fit)
\donttest{plot(resid)}
#qq(dfrr_fit)
}
}
