% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.dfrr.R
\name{summary.dfrr}
\alias{summary.dfrr}
\title{Summary for a dfrr fit}
\usage{
\method{summary}{dfrr}(dfrr_fit)
}
\arguments{
\item{dfrr_fit}{a fitted \code{dfrr}-object obtained from invoking the function \code{\link{dfrr}}.}
}
\description{
Summarise a fitted \code{dfrr}-object. Not implemented.
}
